Add: Controller support for game (D-Pad usage - no left joystick axis support)
Currently the 3.2.X and 3.3.X branches of Godot seem to have issues having multiple actions set to the same input.
https://github.com/godotengine/godot/issues/45628 (issue)
When a joystick axis and a dpad direction are put together the added joystick axis breaks the dpad behaviour. Resulting in the dpad
stoping randmonly after a couple of presses (ranges from 1 to 6 presses before stopping fully).
4.0 should have a fix for this review support then or if the fix is backported to the >=3.3.X series of Godot versions.
https://github.com/godotengine/godot/pull/47599 (fix)